refactor(curator): split inspector into cohesive modules
